pyproject.toml
[project]
name = "a2-cnn-image-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "torch>=2.5.1",
    "torchinfo>=1.8.0",
    "torchvision>=0.20.1",
    "tqdm>=4.67.0",
]

[tool.ruff]
lint.ignore = ["D"]  # Ignore all docstring warnings
